extracted player fire
