+++ src/objects/Player.js
@@ -1,6 +1,10 @@
+var bulletTime = 0;
+var fireButton;
+
 class Player {
 
 	constructor(game){
+		this.game = game;
 		this.player = game.add.sprite(game.world.centerX, game.world.centerY + 200, 'player');
 		game.physics.arcade.enable(this.player);
 	};    
@@ -9,6 +13,7 @@ class Player {
 		this.player.body.collideWorldBounds = true;
 		this.player.animations.add('left', [0, 1, 2, 3, 4], 15, true);
 		this.player.animations.add('right', [0, 1, 2, 3, 4], 15, true);
+		fireButton = this.game.input.keyboard.addKey(Phaser.Keyboard.SPACEBAR);
 	}
 
 	isAlive(player){
@@ -18,6 +23,26 @@ class Player {
 
 	get ship(){
 		return this.player;
+	}
+
+	fire(bullets, laser) {
+	  if(fireButton.isDown) {
+	    this.fireBullet(bullets);
+	    laser.play();
+	    return true;
+	  }
+	}
+
+	fireBullet(bullets) {
+	  if(this.game.time.now > bulletTime){
+	    var bullet = bullets.getFirstExists(false);
+
+	    if(bullet){
+	      bullet.reset(this.player.x + 14, this.player.y);
+	      bullet.body.velocity.y = -400;
+	      bulletTime = this.game.time.now + 200;
+	    }
+	  }
 	}
 }
